% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_geometry.R
\name{as_sf}
\alias{as_sf}
\title{Convert a dataframe to a sf point object}
\usage{
as_sf(data)
}
\arguments{
\item{data}{dataframe with x and y coordinates}
}
\value{
a sf point object, NULL if input dataframe is empty
}
\description{
Convert a dataframe to a sf point object
}
