% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_geometry.R
\name{as_sf_polygon}
\alias{as_sf_polygon}
\title{Convert a sf point object to sf polygon}
\usage{
as_sf_polygon(data)
}
\arguments{
\item{data}{sf point object}
}
\value{
a sf polygon object
}
\description{
Convert a sf point object to sf polygon
}
