% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system_gis.R
\name{as_xyz}
\alias{as_xyz}
\title{Convert a SpatRaster to an xyz tibble}
\usage{
as_xyz(data, z = "z")
}
\arguments{
\item{data}{SpatRaster}

\item{z}{column name for the raster values (default "z")}
}
\value{
tibble with x, y, and named z column
}
\description{
Convert a SpatRaster to an xyz tibble
}
