% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system_attractors.R
\name{bin_xy}
\alias{bin_xy}
\title{Bin xy coordinates to reduce resolution}
\usage{
bin_xy(data, gridsize = 20)
}
\arguments{
\item{data}{dataframe with x and y coordinates}

\item{gridsize}{size of the grid (larger is more precise)}
}
\value{
a downsampled dataframe with x and y coordinates
}
\description{
Bin xy coordinates to reduce resolution
}
