% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_geometry.R
\name{buffer_rectangle}
\alias{buffer_rectangle}
\title{Define a rectangle bounding box with a given ratio around a location}
\usage{
buffer_rectangle(
  data,
  size = 60000,
  x_shift = 0,
  y_shift = 0,
  orientation = "h",
  ratio = 297/210,
  ...
)
}
\arguments{
\item{data}{sf point object}

\item{size}{size of the largest dimension of the bounding box (units)}

\item{x_shift, y_shift}{shifts in the x and y direction (units)}

\item{orientation}{orientation of the bounding box ("h" or "v")}

\item{ratio}{height:width ratio}

\item{...}{used for mapping}
}
\value{
a sf bounding box object
}
\description{
Define a rectangle bounding box with a given ratio around a location
}
