% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_curves.R
\name{f_exp}
\alias{f_exp}
\title{Exponential function}
\usage{
f_exp(x, k = 1, a = 1, b = 0)
}
\arguments{
\item{x}{numeric vector for input values}

\item{k, a, b}{exponential rate, scaling, and asympotic parameters}
}
\value{
a numeric vector
}
\description{
Exponential function
}
