% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_curves.R
\name{f_sig}
\alias{f_sig}
\title{Sigmoid function}
\usage{
f_sig(x, k = 2, a = -2, b = 1, scale = FALSE)
}
\arguments{
\item{x}{numeric vector for input values}

\item{k, a, b}{rate, scaling, and asymptotic parameters}

\item{scale}{Boolean, if TRUE x input values are scaled (scales::rescale)}
}
\description{
Sigmoid function
}
