% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system_gis.R
\name{filter_ridge_slope}
\alias{filter_ridge_slope}
\title{Filter ridgeline dataframe as a function of local slope}
\usage{
filter_ridge_slope(data, size = 3)
}
\arguments{
\item{data}{dataframe of processed ridgelines with zn, xn, y_rank, x_rank columns}

\item{size}{size of window used to compute the slope rolling average (integer, cells)}
}
\value{
filtered dataframe with z_slope column and flat segments set to NA
}
\description{
Computes a rolling average of the absolute slope along each ridge line and
sets flat segments (slope = 0) to NA.
}
