% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system_splines.R
\name{gen_charmap}
\alias{gen_charmap}
\title{Define a map between characters and generated glyphs}
\usage{
gen_charmap(
  seed,
  n = 26,
  n_control = 4,
  n_tall = 4,
  size_tall = 4,
  n_merge = 3,
  n_variation = 10,
  jitter = 1/5,
  scale = 0.5,
  rotation = -pi/6
)
}
\arguments{
\item{seed}{random seed for the character map.}

\item{n}{number of letters in the character map. The total length of the character map is n + 3 (addition of ".", ",", "?")}

\item{n_control}{number of control points for each glyph}

\item{n_tall}{number of tall glyphs}

\item{size_tall}{radius of the set of control points for tall glyphs}

\item{n_merge}{number of glyphs concatenated in the modified map}

\item{n_variation}{number of variations of each glyphs in the map}

\item{jitter}{amount of jitter added to each variations}

\item{scale, rotation}{scale (0-1) and rotation (radian) of the set of control points}
}
\value{
a list-column dataframe with a character and layout column (set of control points).
}
\description{
Generate a set of control points for a collection of glyphs mapped to characters. This initial map is then modified by merging n glyphs into one and adding jitter.
}
