% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system_ode.R
\name{gen_field}
\alias{gen_field}
\title{Simulate an ODE system with initial conditions sampled from a LHS design}
\usage{
gen_field(p, n, t, s = 1/100, x0 = c(-1, 1), y0 = c(0.5, 1))
}
\arguments{
\item{p}{parameter vector for the ODE system}

\item{n}{number of particles to be simulated}

\item{t, s}{end time and step of the simulation (numeric).}

\item{x0, y0}{scaling parameters for the initial conditions}
}
\description{
Simulate an ODE system with initial conditions sampled from a LHS design
}
