% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system_sequence.R
\name{gen_leaf}
\alias{gen_leaf}
\title{Generate a path suitable to represent a leaf-like structure}
\usage{
gen_leaf(i, a = 20, x0 = 0, y0 = 0, shape = "spiral")
}
\arguments{
\item{i}{starting value for the sequence calculation}

\item{a}{value of the angle between leaf segments (degrees)}

\item{x0, y0}{coordinates of the first leaf segments}

\item{shape}{method for the calculation of successive angles between leaf
segments (character). "spiral" accumulates angle in the same direction.
"wave" accumulates angle depending on the parity of the value in the
sequence.}
}
\value{
a dataframe with coordinates of leaf segments
}
\description{
Generate a path suitable to represent a leaf-like structure
}
