% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system_sequence.R
\name{gen_node}
\alias{gen_node}
\title{Generate a collection of path suitable to represent a node-like structure}
\usage{
gen_node(
  n = 20,
  imin = 20,
  imax = 70,
  lmax = 1000,
  end = 1,
  amin = -20,
  amax = 20,
  shift = c(0, 20),
  width = c(0, 15),
  scale = 1,
  shape = "spiral",
  method = "polygon",
  seed = NULL,
  ...
)
}
\arguments{
\item{n}{number of path to simulate}

\item{imin, imax}{bounds of uniform distribution of the sequence starting
value (int)}

\item{lmax}{maximum value for simulated path length}

\item{end}{value used to end path recursion}

\item{amin, amax}{bounds of uniform distribution of the angle between path
segments (degree)}

\item{shift}{a numeric vector for x and y shifts between paths}

\item{width}{a numeric vector for x and y shifts to create polygons from paths}

\item{scale}{scaling value applied on the complete node}

\item{shape}{method for the calculation of successive angles between leaf
elements (character).
\itemize{
\item "spiral" accumulates angle in the same direction.
\item "wave" accumulates angle depending on the parity of the value in the
sequence.
}}

\item{method}{method used to render the path.
\itemize{
\item "spline" fits a b-spline to smooth the initial path.
\item "path" shift the initial path according to \code{width} argument.
\item "polygon" builds an oriented polygon from the shifted path.
\item "polygon_lm" builds an oriented polygon with decreasing size along path length.
}}

\item{seed}{value of the random seed, random if missing}

\item{...}{used for parallel mapping}
}
\value{
a dataframe with coordinates of multiple leafs
}
\description{
Generate a collection of path suitable to represent a node-like structure
}
