% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_rendering.R
\name{gen_paper}
\alias{gen_paper}
\title{Generate a paper texture based on small fibers}
\usage{
gen_paper(
  n = 30000,
  scale = c(1, 1),
  jitter = 1/30,
  strength = 0.5,
  size = 1/10,
  color = "black",
  alpha = 1/20,
  expansion = -1/50
)
}
\arguments{
\item{n}{texture density}

\item{scale}{orientation of texture elements (upper limit of uniform sampling)}

\item{jitter}{length of texture elements}

\item{strength}{curvature in texture elements}

\item{size, color, alpha}{aesthetics of texture elements}

\item{expansion}{additive scaling for limits}
}
\value{
a ggplot object
}
\description{
Generate a paper texture based on small fibers
}
