% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system_splines.R
\name{gen_sequence}
\alias{gen_sequence}
\title{Generate plausible text sequences}
\usage{
gen_sequence(
  seed = NULL,
  method = "lipsum",
  scale = "sentence",
  length = 1,
  size = 3:6
)
}
\arguments{
\item{seed}{random seed for the character sequence.}

\item{method}{method used for sequence generation
\itemize{
\item lipsum, generate latin-based text, with the option to subset the text into smaller elements (scale argument).
\item sentence, draw sentences from the Revised List of Phonetically Balanced Sentences \href{https://www.cs.columbia.edu/~hgs/audio/harvard.html}{Harvard Sentences}
\item random, generate random words within a given size range
}}

\item{scale}{scale of the returned elements, for the lipsum method. Control how the generated text is divided into word, sentence, or text.}

\item{length, }{number of returned elements.}

\item{size}{word size range, for the random method}
}
\value{
a character string.
}
\description{
Generate plausible text sequences
}
