% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system_gis.R
\name{geom_waterline}
\alias{geom_waterline}
\title{Iterate to create geom_sf layers as a function of a dataframe}
\usage{
geom_waterline(data, n = 5, r = 2, d0 = 50, color = "black", scale = 1, ...)
}
\arguments{
\item{data}{sf multipolygon object for shore, used to compute waterlines.}

\item{n}{number of waterlines.}

\item{d0, r}{buffer and rate of buffer growth between water lines.}

\item{color, scale}{color and scaling of plotted lines.}

\item{...}{used for pmap compatibility}
}
\description{
Iterate to create geom_sf layers as a function of a dataframe
}
