% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_geometry.R
\name{get_box}
\alias{get_box}
\title{Get bounding box from 2D objects}
\usage{
get_box(data, method = "center")
}
\arguments{
\item{data}{dataframe with x and y coordinates}

\item{method}{type of output :
\itemize{
\item center : returns the coordinates of center, width and height of the bounding box
}}
}
\value{
a list with bounding box attributes
}
\description{
Get bounding box from 2D objects
}
