% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system_attractors.R
\name{henon_phase}
\alias{henon_phase}
\title{Difference equation for the Henon model}
\usage{
henon_phase(p, xn, yn)
}
\arguments{
\item{p}{parameter vector}

\item{xn, yn}{x and y value for step n}
}
\value{
a vector of length two with n+1 values
}
\description{
Difference equation for the Henon model
}
\references{
http://www.complexification.net/gallery/machines/henonPhase/
}
