% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_partition.R
\name{index_group}
\alias{index_group}
\title{Create group index by partitioning a vector.}
\usage{
index_group(x, k = NULL, length = NULL, method)
}
\arguments{
\item{x}{vector to index}

\item{k, }{number of group, integer}

\item{length, }{length of each group}

\item{method}{discretization method.
\itemize{
\item fixed, fixed size groups, remainder creates an incomplete group
\item quantiles, breaks are defined by quantile probabilities as a function of group number and noise
\item random, breaks are randomly sampled in the vector range as a function of group number.
}}
}
\value{
a vector of indices of the same length as x
}
\description{
Create group index by partitioning a vector.
}
