% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system_attractors.R
\name{iterate}
\alias{iterate}
\title{Iterate difference equations}
\usage{
iterate(f, p, x0, y0, iterations, ...)
}
\arguments{
\item{f}{function to iterate}

\item{p}{parameter vector}

\item{x0, y0}{initial values for x and y}

\item{iterations}{number of iterations}

\item{...}{used for mapping}
}
\value{
a dataframe with x, y coordinates, and iteration number.
}
\description{
Iterate difference equations
}
\references{
https://blog.k2h.se/post/hunting-for-attractors/
}
