% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system_attractors.R
\name{L}
\alias{L}
\title{Compute the Lyapunov exponent to estimate occurrence of a chaotic behaviour}
\usage{
L(f, p, x0, y0, iterations = 1000)
}
\arguments{
\item{f}{function to iterate}

\item{p}{parameter vector}

\item{x0, y0}{initial values for x and y}

\item{iterations}{number of iterations}
}
\value{
The value of the Lyapunov exponent
}
\description{
Compute the Lyapunov exponent to estimate occurrence of a chaotic behaviour
}
\references{
https://blog.k2h.se/post/hunting-for-attractors/
}
