% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_sampling.R
\name{layout_ellipse}
\alias{layout_ellipse}
\title{Sample n points in an elliptic area}
\usage{
layout_ellipse(n = 7, x0 = 0, y0 = 0, r = 1, a = -pi/6, scale_x = 0.5)
}
\arguments{
\item{n}{number of points}

\item{x0, y0}{coordinates of the center of the sampling area}

\item{r}{radius of the sampling area}

\item{a}{rotation angle of the point set (radians)}

\item{scale_x}{scaling coefficient applied on the x axis}
}
\value{
a dataframe with n, x, and y columns
}
\description{
Point are sampled from uniform distribution in polar coordinates. The point set is then transformed to cartesian coordinates, scaled, and rotated.
}
