% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system_splines.R
\name{layout_glyph}
\alias{layout_glyph}
\title{Create a concatenated set of control points (glyph) from a sequence of characters.}
\usage{
layout_glyph(seq, map, shift = 1)
}
\arguments{
\item{seq}{a sequence of characters (string)}

\item{map}{a character map generated by \code{gen_charmap()}}

\item{shift}{horizontal shift when concatenating individual sets of control points}
}
\value{
a dataframe of concatenated sets of control points
}
\description{
One glyph is created from n characters, defined by individual sets of control points.
}
