% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_sampling.R
\name{layout_rectangle}
\alias{layout_rectangle}
\title{Sample n points uniformly in a rectangular area}
\usage{
layout_rectangle(n = 7, method = "uniform", xlim = c(-1, 1), ylim = c(-1, 1))
}
\arguments{
\item{n}{number of points}

\item{method}{method used to sample the space
\itemize{
\item "uniform": uniform distributions for each dimensions.
\item "Halton": Halton low discrepancy sequence from \code{randtoolbox::halton()}.
}}

\item{xlim, ylim}{x and y limits, default to (-1,1)}
}
\value{
a dataframe with n, x, and y columns
}
\description{
Sample n points uniformly in a rectangular area
}
