% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system_splines.R
\name{layout_word}
\alias{layout_word}
\title{Create a concatenated set of control points from a sequence of characters.}
\usage{
layout_word(word, data, shift = 1)
}
\arguments{
\item{word}{a sequence of characters (string)}

\item{data}{a character map generated by \code{gen_charmap()}}

\item{shift}{horizontal shift when concatenating individual sets of control points}
}
\value{
a dataframe of concatenated sets of control points
}
\description{
One glyph is created from n characters, defined by individual sets of control points.
}
