% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system_attractors.R
\name{logistic_map}
\alias{logistic_map}
\title{Difference equation for the logistic model}
\usage{
logistic_map(p, xn, yn)
}
\arguments{
\item{p}{parameter vector}

\item{xn, yn}{x and y value for step n}
}
\value{
a vector of length two with n+1 values
}
\description{
Difference equation for the logistic model
}
