% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system_attractors.R
\name{normalize_xy}
\alias{normalize_xy}
\title{Normalize xy coordinates}
\usage{
normalize_xy(data, n = 100)
}
\arguments{
\item{data}{dataframe with x and y coordinates}

\item{n}{number of starting iterations to remove}
}
\value{
a normalized dataframe with x and y coordinates
}
\description{
Normalize xy coordinates
}
