% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_partition.R
\name{partition}
\alias{partition}
\title{Partition a unit into k parts working with 1/n th fractions combinations}
\usage{
partition(k = 1:3, n = 6, sample = TRUE)
}
\arguments{
\item{k, }{number of parts (1:5), vector}

\item{n, }{partition fraction (1-5), integer}

\item{sample, }{if TRUE (default) sample a random partition fitting the conditions,
else enumerate distinct partitions}
}
\value{
a random vector of fraction summing to 1
}
\description{
Partition a unit into k parts working with 1/n th fractions combinations
}
