% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system_attractors.R
\name{quadratic_map}
\alias{quadratic_map}
\title{Difference equations for quadratic model}
\usage{
quadratic_map(p, xn, yn)
}
\arguments{
\item{p}{parameter vector}

\item{xn, yn}{x and y value for step n}
}
\value{
a vector of length two with n+1 values
}
\description{
Difference equations for quadratic model
}
\references{
https://blog.k2h.se/post/hunting-for-attractors/
}
