% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_geometry.R
\name{r_t}
\alias{r_t}
\title{Rotate then translate a 2D object}
\usage{
r_t(data, x0, y0, a)
}
\arguments{
\item{data}{dataframe with x and y coordinates}

\item{x0, y0}{shift in x and y axis (units)}

\item{a}{rotation angle (radian)}
}
\value{
a dataframe with new coordinates columns
}
\description{
Rotate then translate a 2D object
}
