% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system_gis.R
\name{read_tiles}
\alias{read_tiles}
\title{Read DEM tiles from IGN public elevation datasets}
\usage{
read_tiles(coord, buffer, source)
}
\arguments{
\item{coord}{sf point in EPSG:2154, or named vector with lon/lat in decimal degrees}

\item{buffer}{circular buffer around the centroid (numeric, m).}

\item{source}{DEM database to be used (character).
\itemize{
\item "db_alti" is a national database at 25x25m resolution.
\item "rge_alti" is a gridded database at 5x5m resolution, with 1 - 20m source data.
}}
}
\value{
stars object
}
\description{
Read DEM tiles from IGN public elevation datasets
}
