% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system_gis.R
\name{render_contour}
\alias{render_contour}
\title{Render a DEM as contour plot with waterlines}
\usage{
render_contour(
  data,
  coord,
  area_min = 20000,
  length_min = 600,
  alt_min,
  alt_max,
  alt_contour = 100,
  n_water = 5,
  scale = 1,
  outline = FALSE,
  layers = FALSE,
  ...
)
}
\arguments{
\item{data}{a digital elevation model as a spatial data object (stars) from \code{read_dem()}}

\item{coord}{geographical coordinates of the centroid (named vector with lon, lat)}

\item{area_min}{minimal area for polygons to be considered (m^2)}

\item{length_min}{minimal length for contour lines to be considered (m)}

\item{alt_min, alt_max, alt_contour}{minimal, maximal, and elevation step used to define contour lines (m)}

\item{n_water}{number of waterlines}

\item{scale}{scaling coefficient of contour and water lines}

\item{outline}{if TRUE render and plot only shorelines (boolean)}

\item{layers}{if TRUE return all layers in different objects (boolean)}

\item{...}{used for pmap compatibility}
}
\value{
a ggplot object
}
\description{
Render a DEM as contour plot with waterlines
}
