% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system_graph.R
\name{render_graph}
\alias{render_graph}
\title{Render point layouts with different graphs and aesthetics}
\usage{
render_graph(
  data,
  vars = c(x, y),
  graph = "rng",
  aes = "default",
  k = 3,
  n = 15,
  strength = 0.5,
  color = "black",
  width = 0.5,
  alpha = 1,
  coord = NULL
)
}
\arguments{
\item{data}{a dataframe of 2D coordinates}

\item{vars}{names of the coordinates to be used (unquoted vector)}

\item{graph}{method used to compute the graph object from the point layout.
\itemize{
\item "rng": \href{https://en.wikipedia.org/wiki/Relative_neighborhood_graph}{relative neighborhood graph},
roughly defines the human perception of the shape of the point set.
\item "knn": \href{https://en.wikipedia.org/wiki/Nearest_neighbor_graph}{nearest neighbor graph}, connect k adjacent points together.,
\item "mst" : \href{https://en.wikipedia.org/wiki/Minimum_spanning_tree}{minimum spanning tree}, connect all vertices with the minimal edge length.
}}

\item{aes}{aesthetic of the rendered plot.
\itemize{
\item "default": point and lines.
\item "line": only lines.
\item "bezier": Draw edges as Bezier curves.
\item "arc" : Draw edges as Arcs
}}

\item{k}{number of neighbors to account to in both methods. NA is used to calculate the true relative neighborhood graph but increase to compute time.}

\item{n}{number of points created along the paths}

\item{strength}{strength of the curvature of the bend.}

\item{color, width, alpha}{fixed aesthetic parameters for the ggplot object}

\item{coord}{ggplot2 coordinate system object passed to ggplot}
}
\value{
a ggplot object
}
\description{
Render point layouts with different graphs and aesthetics
}
