% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system_splines.R
\name{render_line}
\alias{render_line}
\title{Render a line of glyphs}
\usage{
render_line(data, ncol = 80, scale = 0.9)
}
\arguments{
\item{data}{a list of glyphs as ggplot objects}

\item{ncol}{maximum number of glyphs on a line}

\item{scale}{scaling parameter for individual glyphs}
}
\description{
Render a line of glyphs
}
