% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system_sequence.R
\name{render_node}
\alias{render_node}
\title{Render nodes with different aesthetic}
\usage{
render_node(
  data,
  method = "polygon",
  radius = 0,
  xlim = NULL,
  ylim = NULL,
  margin = 0
)
}
\arguments{
\item{data}{a dataframe of objects coordinates produced by \code{gen_node()}
function}

\item{method}{method used to render objects. "spline" render paths using
B-splines. "segment" render leaf as a polygon with border and visible
segments. "polygon" render closed path using \code{geom_shape()}.}

\item{radius}{radius of polygon smoothing}

\item{xlim, ylim}{limits passed to \code{coord_fixed()}}

\item{margin}{margins passed to \code{theme()}}
}
\value{
a ggplot object
}
\description{
Render nodes with different aesthetic
}
