% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system_attractors.R
\name{render_paragraph}
\alias{render_paragraph}
\title{Render a grid of plot using a text sequence as pattern}
\usage{
render_paragraph(text, data, ncol = 80, scale = 0.8, align = "none")
}
\arguments{
\item{text}{a character string}

\item{data}{a set of plot of characters}

\item{ncol}{number of columns of the grid of plots}

\item{scale}{scale of individual plots}

\item{align}{parameter for plot alignment (cowplot)}
}
\value{
a grid of plots
}
\description{
Render a grid of plot using a text sequence as pattern
}
