% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system_attractors.R
\name{render_plot}
\alias{render_plot}
\title{Render a single character/scatterplot}
\usage{
render_plot(data, color = "black", size = 0.5, alpha = 1/10)
}
\arguments{
\item{data}{dataframe with x and y coordinates}

\item{color}{Point color}

\item{size}{Point size}

\item{alpha}{Point alpha}
}
\value{
a plot for a single point set
}
\description{
Render a single character/scatterplot
}
