% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system_gis.R
\name{render_ridge}
\alias{render_ridge}
\title{Render a DEM with elevation as a function of longitude, grouped per latitude (ridge plot)}
\usage{
render_ridge(
  data,
  n_ridges = 200,
  n_drop = 0,
  n_lag = 100,
  z_shift = 15,
  z_threshold = 10,
  z_k = 0
)
}
\arguments{
\item{data}{dataframe from a digital elevation model (x,y,z)}

\item{n_ridges}{number of ridges uniformly selected, value 0 select all ridges in dataset  (integer)}

\item{n_drop}{number of ridges to drop in distance (integer)}

\item{n_lag}{depth of search for line removal algorithm (number of successive ridges)}

\item{z_shift}{distance on y-axis used to shift successive ridges (m)}

\item{z_threshold}{minimal distance threshold to remove points
between successive ridges (m)}

\item{z_k}{scaling coefficient for perspective computation. a value of 0 switch to linear perspective.}
}
\value{
a dataframe suitable for plotting in two dimensions :
\itemize{
\item x, longitude (m)
\item y, latitude (m)
\item z, altitude (m)
\item xn, relative longitude (m)
\item y_rank, relative latitude
\item y_dist, relative latitude (0,1)
\item dz, shift in altitude (m)
\item zs, shifted altitude (m)
\item z_rank, relative altitude (ranked per longitude)
\item zn, shifted altitude, after occlusion (m)
\item zl_n, length of the current ridge (cell)
}
}
\description{
Render a DEM with elevation as a function of longitude, grouped per latitude (ridge plot)
}
