% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system_splines.R
\name{render_script}
\alias{render_script}
\title{Render a glyph set as a function of a character string and a character map.}
\usage{
render_script(
  text,
  data,
  length = 80,
  scale = 0.9,
  orientation = "h",
  align = "none"
)
}
\arguments{
\item{text}{a long character sequence}

\item{data}{a character map generated by \code{gen_charmap()}}

\item{length}{maximum number of glyphs on a line or columns}

\item{scale}{scaling parameter for individual glyphs}

\item{orientation}{option to rotate the point set 90° clockwise}

\item{align}{specifies whether graphs in the grid should be horizontally ("h") or vertically ("v") aligned.}
}
\value{
a ggplot object.
}
\description{
Emulates script-like writing by juxtaposing glyphs, and adding individual variation.
}
