% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system_attractors.R
\name{render_sequence}
\alias{render_sequence}
\title{Sample and render a sequence of plots in a list}
\usage{
render_sequence(n, length, data)
}
\arguments{
\item{n}{size of the sample to draw}

\item{length}{size of the grid used to draw individual plots}

\item{data}{vector of plots to sample}
}
\value{
a grid of plots
}
\description{
Sample and render a sequence of plots in a list
}
