% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system_splines.R
\name{render_spline}
\alias{render_spline}
\title{Render spline curves from control points and grouping indices.}
\usage{
render_spline(
  data,
  type = "clamped",
  n = 100,
  color = "black",
  width = 0.5,
  alpha = 1,
  coord = ggplot2::coord_fixed()
)
}
\arguments{
\item{data}{a dataframe with x and y columns defining control points, and \emph{group} and \emph{glyph} columns defining grouping structure.}

\item{type}{Either 'clamped' (default) or 'open'. Ensures the spline starts and ends at the terminal control points.}

\item{n}{number of points generated for the spline curve.}

\item{color, width, alpha}{arguments passed to \code{geom_bspline()}}

\item{coord}{coordinate system passed to \code{ggplot()}, default to coord_fixed()}
}
\value{
a ggplot object
}
\description{
Render spline curves from control points and grouping indices.
}
