% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system_splines.R
\name{render_text}
\alias{render_text}
\title{Render text from glyph geometry.}
\usage{
render_text(
  data,
  method = "spline",
  type = "clamped",
  n = 100,
  color = "black",
  width = 0.5,
  alpha = 1,
  coord = ggplot2::coord_fixed()
)
}
\arguments{
\item{data}{a dataframe with x and y columns, and \emph{group} and \emph{glyph} columns defining grouping structure.}

\item{method}{rendering method: "spline" (B-spline interpolation) or "path" (straight line segments)}

\item{type}{Either 'clamped' (default) or 'open'. Only used with method = "spline".}

\item{n}{number of points generated for the spline curve. Only used with method = "spline".}

\item{color, width, alpha}{aesthetics passed to the geom}

\item{coord}{coordinate system passed to \code{ggplot()}, default to coord_fixed()}
}
\value{
a ggplot object
}
\description{
Render text from glyph geometry.
}
