% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system_gis.R
\name{roll_max_lag}
\alias{roll_max_lag}
\title{Sliding maximum of the n previous values}
\usage{
roll_max_lag(x, n)
}
\arguments{
\item{x}{numeric vector}

\item{n}{window size (number of previous values to look back)}
}
\description{
Pads x with n values of -Inf so roll_max always has a full window at group
boundaries (-Inf is the identity element for max). Returns a vector of same length as x.
}
