% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_geometry.R
\name{rotate}
\alias{rotate}
\title{Rotate a 2D object}
\usage{
rotate(data, a)
}
\arguments{
\item{data}{dataframe with x and y coordinates}

\item{a}{rotation angle (radian)}
}
\value{
a dataframe with new coordinates columns
}
\description{
Rotate a 2D object
}
