% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_partition.R
\name{round_near}
\alias{round_near}
\title{Round down the maximum value of x to the nearest multiple of accuracy}
\usage{
round_near(x, accuracy)
}
\arguments{
\item{x, }{numeric value or vector}

\item{accuracy, }{integer number defining the rounding limit}
}
\value{
integer value
}
\description{
Round down the maximum value of x to the nearest multiple of accuracy
}
