% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_sampling.R
\name{sample_box}
\alias{sample_box}
\title{Sample a rectangular area defined by its relative size and aspect ratio nested in a larger area.}
\usage{
sample_box(xlim = c(-1, 1), ylim = c(-1, 1), r = 0.7, p = 0.2)
}
\arguments{
\item{xlim, ylim}{limits of the larger sampling area}

\item{r}{aspect ratio of the small sampled area}

\item{p}{proportion of the small sampled area (relative length)}
}
\description{
Sample a rectangular area defined by its relative size and aspect ratio nested in a larger area.
}
