% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_sampling.R
\name{sample_disc}
\alias{sample_disc}
\title{Sample n points evenly distributed on a disc}
\usage{
sample_disc(n = 100, x0 = 0, y0 = 0, r = 1, a = pi * (1 + sqrt(5)))
}
\arguments{
\item{n}{number of points}

\item{x0, y0}{coordinates of the center of the sampled circle}

\item{r}{radius of the sampled circle}

\item{a}{angle in polar coordinates (radians)}
}
\value{
a dataframe with n, x, and y columns
}
\description{
Points are arranged using Fermat's spiral (\href{https://doi.org/10.1016\%2F0025-5564\%2879\%2990080-4}{Vogel, 1979})
}
