% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system_splines.R
\name{sample_letters}
\alias{sample_letters}
\title{Sample n letters in the R letters set.}
\usage{
sample_letters(n = 3, x = letters)
}
\arguments{
\item{n}{number of sampled elements}

\item{x}{set of letters to sample from}
}
\value{
a concatenated character string
}
\description{
Sample n letters in the R letters set.
}
