% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_sampling.R
\name{sample_missing}
\alias{sample_missing}
\title{Sample and replace values with NA in a vector}
\usage{
sample_missing(x, y = NULL, p, method = "random")
}
\arguments{
\item{x, y}{input vectors}

\item{p}{proportion of missing values (0,1)}

\item{method}{method from sampling values}
}
\value{
a vector with missing values.
}
\description{
Sample and replace values with NA in a vector
}
