% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_sampling.R
\name{sample_perimeter}
\alias{sample_perimeter}
\title{Sample n point regularly spaced on a circle}
\usage{
sample_perimeter(n = 7, x0 = 0, y0 = 0, r = 1)
}
\arguments{
\item{n}{number of points}

\item{x0, y0}{coordinates of the center of the sampled circle}

\item{r}{radius of the sampled circle}
}
\value{
a dataframe with n, x, and y columns
}
\description{
Sample n point regularly spaced on a circle
}
