% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_sampling.R
\name{sample_rectangle}
\alias{sample_rectangle}
\title{Sample n point regularly spaced on a rectangle}
\usage{
sample_rectangle(n = 200, scaling = 0, jitter = 10, x0, y0, x, y)
}
\arguments{
\item{n}{number of points}

\item{scaling}{expansion factor}

\item{jitter}{jitter factor}

\item{x0, y0, x, y}{coordinates of the center, width and height of the sampled rectangle}
}
\value{
a dataframe with x and y columns
}
\description{
Sample n point regularly spaced on a rectangle
}
