% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system_attractors.R
\name{sample_sequence}
\alias{sample_sequence}
\title{Random sampling of character sequence}
\usage{
sample_sequence(set = LETTERS[1:25], length = 12)
}
\arguments{
\item{set}{Letter set to use}

\item{length}{Length of the output sequence}
}
\value{
a random character sequence of given length
}
\description{
Random sampling of character sequence
}
