% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system_gis.R
\name{sample_slope}
\alias{sample_slope}
\title{Remove points as a function of local slope along ridge lines}
\usage{
sample_slope(data, size = 3)
}
\arguments{
\item{data}{dataframe with zn, xn, y_rank, x_rank columns (from mask_lines)}

\item{size}{window size for the slope rolling average (integer, cells)}
}
\value{
input dataframe with additional column:
\itemize{
\item \code{z_slope}: rolling mean of absolute slope per ridge line
}
}
\description{
Computes a rolling average of the absolute slope along each ridge line and
sets flat segments (slope = 0) to NA, creating line interruptions in geom_path.
}
