% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_sampling.R
\name{sample_vessel}
\alias{sample_vessel}
\title{Sample equidistant points on the perimeter of l concentric circles of r radius}
\usage{
sample_vessel(x0 = 0, y0 = 0, l = 5, r = 1/100, jitter = 1/500)
}
\arguments{
\item{x0, y0}{coordinates of the center of the concentric circles}

\item{l}{number of concentric circles}

\item{r}{radius of the outermost circle}

\item{jitter}{amount of jitter to apply to sampled points}
}
\description{
Sample equidistant points on the perimeter of l concentric circles of r radius
}
