% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_files.R
\name{save_plot}
\alias{save_plot}
\title{Save R plot with presets file formats}
\usage{
save_plot(
  plot,
  file,
  path = "R/figures",
  size = c(148, 210),
  format = "jpg",
  bg = "#FEFAEE"
)
}
\arguments{
\item{plot}{R plot object}

\item{file}{file name without extension}

\item{path}{file path without trailing backslash}

\item{size}{numeric vector for image width and height in mm, default to A5}

\item{format}{file format
\itemize{
\item png, print quality bitmap output (400 dpi, lossless)
\item jpg, default bitmap output
\item snap, screen quality output (150 dpi, default path)
\item svg, vector format
}}

\item{bg}{background color passed to ggsave function}
}
\description{
Save R plot with presets file formats
}
