% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system_sequence.R
\name{seq_alt}
\alias{seq_alt}
\title{Calculate a sequence of alternate values sampled from a Normal distribution}
\usage{
seq_alt(n, m = 60, sd = 10)
}
\arguments{
\item{n}{sequence length}

\item{m}{mean value of the Normal distribution}

\item{sd}{standard deviation of the Normal distribution}
}
\value{
a numeric vector of length n
}
\description{
Calculate a sequence of alternate values sampled from a Normal distribution
}
