% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system_sequence.R
\name{seq_collatz}
\alias{seq_collatz}
\title{Calculate the Collatz sequence recursively}
\usage{
seq_collatz(i, end = 1, max = 1000)
}
\arguments{
\item{i}{initial value}

\item{end}{value used to end recursion}

\item{max}{maximum number of iterations}
}
\value{
a numeric vector
}
\description{
Calculate the Collatz sequence recursively
}
