% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_sampling.R
\name{seq_noise}
\alias{seq_noise}
\title{Create a regular sequence of discrete numbers with noise}
\usage{
seq_noise(start, end, n, jitter = 1)
}
\arguments{
\item{start, end}{bounds for the generated sequence}

\item{n}{sequence length}

\item{jitter}{bounds for the noise distribution}
}
\value{
a vector of integers
}
\description{
Create a regular sequence of discrete numbers with noise
}
