% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_geometry.R
\name{set_range}
\alias{set_range}
\title{Set common limits to a list of plots}
\usage{
set_range(p)
}
\arguments{
\item{p}{a list of plots}
}
\value{
a list of plots with modified limits.
}
\description{
Set common limits to a list of plots
}
