% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system_ode.R
\name{simulate_lorenz}
\alias{simulate_lorenz}
\title{Simulator for the Lorenz dynamical system}
\usage{
simulate_lorenz(
  start_coords = c(x = 1, y = 1, z = 1),
  t_max = 10,
  t_step = 1/100
)
}
\arguments{
\item{start_coords}{initial conditions (numeric vector)}

\item{t_max, t_step}{end time and step of the simulation (numeric).}
}
\value{
a dataframe with time, x, and y columns
}
\description{
Describes a simplified mathematical model for atmospheric convection. The parameters are set so that the system exhibits a chaotic behavior (σ = 10, ρ = 28, β = 8/3).
}
