% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system_ode.R
\name{simulate_lv}
\alias{simulate_lv}
\title{Simulator for the Lotka-Volterra dynamical system}
\usage{
simulate_lv(
  parameters,
  start_coords = c(x = 0.01, y = 0.01),
  t_max = 10,
  t_step = 1/100
)
}
\arguments{
\item{parameters}{α, β, γ, δ are positive real parameters describing the interaction of the two species. (numeric vector).}

\item{start_coords}{initial conditions (numeric vector)}

\item{t_max, t_step}{end time and step of the simulation (numeric).}
}
\value{
a dataframe with time, x, and y columns
}
\description{
Describes the \href{https://en.wikipedia.org/wiki/Lotka\%E2\%80\%93Volterra_equations}{dynamics} of biological systems in which two species interact.
}
