% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system_ode.R
\name{simulate_quadratic}
\alias{simulate_quadratic}
\title{Simulate discrete time quadratic maps}
\usage{
simulate_quadratic(
  parameters,
  start_coords = c(x = 0, y = 0),
  iterations = 100
)
}
\arguments{
\item{parameters}{a numeric vector of 12 values}

\item{start_coords}{initial conditions (numeric vector)}

\item{iterations}{number of iterations (integer)}
}
\value{
a dataframe with time, x, and y columns
}
\description{
Based on the original article by J.C. Sprott (Sprott, J. C. "Automatic Generation of Strange Attractors." Comput. & Graphics 17, 325-332, 1993c). The map used formalized as
\itemize{
\item xn+1 = a0 + a1 xn + a2 xn2 + a3 xn yn + a4 yn + a5 yn2
\item yn+1 = a6 + a7 xn + a8 xn2 + a9 xn yn + a10 yn + a11 yn2
}
}
