% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_geometry.R
\name{tr_jitter}
\alias{tr_jitter}
\title{Add jitter to a 2D object}
\usage{
tr_jitter(data, a, id = FALSE)
}
\arguments{
\item{data}{dataframe with x and y coordinates, and id columns}

\item{a}{jitter amount (units)}

\item{id}{if TRUE add the id column of the input to the result dataframe}
}
\value{
a dataframe with new coordinates columns
}
\description{
Add jitter to a 2D object
}
