% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_geometry.R
\name{tr_loess}
\alias{tr_loess}
\title{smooth paths with loess method}
\usage{
tr_loess(data, span, n_min = 10, ...)
}
\arguments{
\item{data}{a dataframe with x and y columns}

\item{span}{controls the degree of smoothing. higher value creates smoother paths.}

\item{n_min}{minimal number of points for smoothing.}

\item{...}{use for mapping other arguments}
}
\description{
smooth paths with loess method
}
