% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_geometry.R
\name{tr_recurse}
\alias{tr_recurse}
\title{apply recursive transformations}
\usage{
tr_recurse(data, f, p, ...)
}
\arguments{
\item{data}{dataframe with xy columns}

\item{f}{transformation function}

\item{p}{named list of parameters values for the transformation function}

\item{...}{use for mapping other arguments}
}
\description{
apply recursive transformations
}
