% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_geometry.R
\name{tr_rotate}
\alias{tr_rotate}
\title{Rotate a 2D object}
\usage{
tr_rotate(data, a, id = FALSE)
}
\arguments{
\item{data}{dataframe with x and y coordinates, and id columns}

\item{a}{rotation angle (radian)}

\item{id}{if TRUE add the id column of the input to the result dataframe}
}
\value{
a dataframe with new coordinates columns
}
\description{
Rotate a 2D object
}
