% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_geometry.R
\name{tr_wave}
\alias{tr_wave}
\title{Apply a periodic transformation to a 2D object}
\usage{
tr_wave(data, period = 30, amplitude = 1/10, delta = 0)
}
\arguments{
\item{data}{dataframe with x and y coordinates}

\item{period}{period for cyclic offset}

\item{amplitude}{strength of offset}

\item{delta}{control how x and y offsets are linked}
}
\value{
a dataframe with new coordinates columns
}
\description{
Apply a periodic transformation to a 2D object
}
