% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system_sequence.R
\name{transform_path}
\alias{transform_path}
\title{Renders a 2D path with different methods}
\usage{
transform_path(data, scale = 1, width = c(0, 10), method = "polygon")
}
\arguments{
\item{data}{a dataframe with x and y coordinates}

\item{scale}{scaling factor of the output}

\item{width}{a numeric vector for x and y shifts of the path}

\item{method}{method used to render the path. "spline" fits a b-spline to
smooth the initial path. "path" shift the initial path according to \code{width}
argument. "polygon" builds an oriented polygon. "polygon_lm" builds an
oriented polygon with decreasing size along path length.}
}
\value{
a dataframe with new x and y coordinates
}
\description{
Renders a 2D path with different methods
}
