% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system_sequence.R
\name{transform_vector}
\alias{transform_vector}
\title{Recursively transform a vector as a function of length and angle values}
\usage{
transform_vector(.x, .y)
}
\arguments{
\item{.x}{initial or n-1 coordinate vector}

\item{.y}{n coordinate vector}
}
\value{
A vector the same length of .x with the same names as .x
}
\description{
Recursively transform a vector as a function of length and angle values
}
