% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_geometry.R
\name{translate}
\alias{translate}
\title{Translate a 2D object}
\usage{
translate(data, x0, y0)
}
\arguments{
\item{data}{dataframe with x and y coordinates}

\item{x0, y0}{shift in x and y axis (units)}
}
\value{
a dataframe with new coordinates columns
}
\description{
Translate a 2D object
}
