% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system_sequence.R
\name{collatz}
\alias{collatz}
\title{Define rules for calculating the Collatz sequence (with shortcut for odd numbers)}
\usage{
collatz(n)
}
\arguments{
\item{n}{value for step \code{n}}
}
\value{
value for step \code{n + 1}
}
\description{
Define rules for calculating the Collatz sequence (with shortcut for odd numbers)
}
\references{
https://en.wikipedia.org/wiki/Collatz_conjecture
}
