% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_geometry.R
\name{crop_rectangle}
\alias{crop_rectangle}
\title{Crop a raster (stars format) to a given bounding box}
\usage{
crop_rectangle(raster, point, size)
}
\arguments{
\item{raster, }{input raster (stars)}

\item{point, }{centroid of the cropping region. point and raster CRS must be identical (sf).}

\item{size, }{dimensions defining the cropping bounding box (x,y numerical vector)}
}
\value{
xyz dataframe of the cropped region
}
\description{
Crop a raster (stars format) to a given bounding box
}
