% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system_gis.R
\name{filter_ridge_rank}
\alias{filter_ridge_rank}
\title{Filter ridgeline dataframe as a function of rank index}
\usage{
filter_ridge_rank(data, p, method = "grid")
}
\arguments{
\item{data}{dataframe of computed ridgelines from \code{render_ridge()}}

\item{p}{proportion of ridgeline to remove}

\item{method}{method used to remove ridges.
\itemize{
\item "random" randomly removes a proportion of ridges.
\item "grid" uniformly sample a proportion of ridges.
\item "strip" remove a proportion of strips of ridges with a low elevation variance.
}}
}
\description{
Filter ridgeline dataframe as a function of rank index
}
