% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system_gis.R
\name{filter_ridge_slope}
\alias{filter_ridge_slope}
\title{Filter ridgeline dataframe as a function of local slope}
\usage{
filter_ridge_slope(data, size = 3)
}
\arguments{
\item{data}{dataframe of computed ridgelines from render_ridge()}

\item{size}{size of window used to compute the slope rolling average (integer, cells)}
}
\description{
Filter ridgeline dataframe as a function of local slope
}
