% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system_splines.R
\name{gen_charmap}
\alias{gen_charmap}
\title{Define a map between letters and glyphs}
\usage{
gen_charmap(
  n = 26,
  n_control = 4,
  n_tall = 4,
  size_tall = 4,
  scale = 0.5,
  rotation = -pi/6
)
}
\arguments{
\item{n}{number of letters in the character map. the total length of the character map is n + 3 (addition of ".", ",", "?")}

\item{n_control}{number of control points for each glyph}

\item{n_tall}{number of tall glyphs}

\item{size_tall}{radius of the set of control points for tall glyphs}

\item{scale, rotation}{scale (0-1) and rotation (radian) of the set of control points}
}
\value{
a dataframe with a character and layout column (set of control points).
}
\description{
Generate a set of control points for a collection of glyphs mapped to characters
}
