% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_geometry.R
\name{get_box}
\alias{get_box}
\title{Get bounding box from 2D dataframes}
\usage{
get_box(data, vars = c("x", "y"), method = "center")
}
\arguments{
\item{data}{dataframe with x and y coordinates}

\item{vars}{alternative names for x and y variables (character vector)}

\item{method}{type of output :
\itemize{
\item center : returns the coordinates of center, width and height of the bounding box
\item points : returns the coordinates of cardinal points
}}
}
\value{
bounding box attributes
}
\description{
Get bounding box from 2D dataframes
}
