% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system_splines.R
\name{layout_paragraph}
\alias{layout_paragraph}
\title{Layout a glyph set as a function of a character string and a character map.}
\usage{
layout_paragraph(
  seq,
  map,
  cut = 80,
  scale = c(2.5, 5),
  noise = c(0, 0),
  spacing = 1,
  orientation = "lrtb"
)
}
\arguments{
\item{seq}{a character sequence.}

\item{map}{a character map generated by \code{gen_charmap()}}

\item{cut}{maximum number of glyphs on a line or columns}

\item{scale}{scales between glyphs (x) and lines (y)}

\item{noise}{random spacing between characters (sd) and words (mean, sd)}

\item{spacing}{relative space between paragraphs}

\item{orientation}{writing direction (e.g. right to left, top to bottom)}
}
\value{
a dataframe with coordinates of glyphs.
}
\description{
Emulates text by juxtaposing glyphs, and adding individual variation.
}
