% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system_splines.R
\name{layout_sequence}
\alias{layout_sequence}
\title{Layout a glyph set as a function of a character string and a character map.}
\usage{
layout_sequence(seq, map, cut = 80, scale = c(2.5, 5), orientation = "lrtb")
}
\arguments{
\item{seq}{a character sequence.}

\item{map}{a character map generated by \code{gen_charmap()}}

\item{cut}{maximum number of glyphs on a line or columns}

\item{scale}{scaling parameter for individual glyphs}

\item{orientation}{writing direction (e.g. right to left, top to bottom)}
}
\value{
a dataframe with coordinates of glyphs.
}
\description{
Emulates text by juxtaposing glyphs, and adding individual variation.
}
