% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_sampling.R
\name{layout_square}
\alias{layout_square}
\title{Sample n points uniformly in a square}
\usage{
layout_square(
  n = 7,
  method = "uniform",
  xlim = c(-1, 1),
  ylim = c(-1, 1),
  seed
)
}
\arguments{
\item{n}{number of points}

\item{method}{method used to sample the space
\itemize{
\item "uniform": basic sampling based on independent uniform distributions for each dimensions.
\item "sobol": sampling using uniform Sobol low discrepancy sequences from \code{randtoolbox::sobol()}.
}}

\item{xlim, ylim}{x and y limits, default to (-1,1)}

\item{seed}{seed used for the sobol method.}
}
\value{
a dataframe with n, x, and y columns
}
\description{
Sample n points uniformly in a square
}
