% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system_splines.R
\name{render_cursive}
\alias{render_cursive}
\title{Render a glyph set as a function of a character string and a character map.}
\usage{
render_cursive(
  text,
  map,
  ncol = 80,
  shift = 1,
  color = "black",
  size = 0.5,
  n_points = 200,
  scale = 0.8
)
}
\arguments{
\item{text}{a long character sequence}

\item{map}{a character map generated by \code{gen_charmap()}}

\item{ncol}{maximum number of characters per line.}

\item{shift}{horizontal shift when concatenating individual sets of control points}

\item{color, size}{color and size of the spline used to draw glyphs}

\item{n_points}{number of points generated for the spline}

\item{scale}{scaling parameter applied on individual lines fo glyphs}
}
\value{
a ggplot object
}
\description{
Emulates cursive-like writing by concatenating glyphs.
}
