% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_geometry.R
\name{render_frame}
\alias{render_frame}
\title{Render a rectangular frame around a 2D object.}
\usage{
render_frame(data, jitter = 10, scaling = 0.1, size = 0.5, color = "black")
}
\arguments{
\item{data}{data frame with x and y columns, it could be the plotted data or just the limits for the frame.}

\item{jitter}{amount of jitter to add to sampled points in the frame}

\item{scaling}{scaling factor around the object bounding box}

\item{size, color}{aesthetics passed to geom_path()}
}
\value{
a geom_path ggplot layer.
}
\description{
Render a rectangular frame around a 2D object.
}
