% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system_splines.R
\name{render_script}
\alias{render_script}
\title{Render a glyph set as a function of a character string and a character map.}
\usage{
render_script(
  text,
  map,
  length = 80,
  scale = c(2.5, 5),
  n = 100,
  orientation = "lrtb",
  coord = NULL,
  color = "black",
  size = 0.3
)
}
\arguments{
\item{text}{a long character sequence}

\item{map}{a character map generated by \code{gen_charmap()}}

\item{length}{maximum number of glyphs on a line or columns}

\item{scale}{scaling parameter for individual glyphs}

\item{n}{number of data points generated for the spline}

\item{orientation}{writing direction (e.g. right to left, top to bottom)}

\item{coord}{coordinate system for the output ggplot object}

\item{color}{text color}

\item{size}{text linewidth}
}
\value{
a ggplot object.
}
\description{
Emulates script-like writing by juxtaposing glyphs, and adding individual variation.
}
