% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system_splines.R
\name{render_spline}
\alias{render_spline}
\title{Render a spline curve from control points}
\usage{
render_spline(
  data,
  type = "clamped",
  n = 100,
  orientation = "h",
  width = 0.5,
  alpha = 1,
  coord = NULL
)
}
\arguments{
\item{data}{a dataframe with x and y columns defining control points.}

\item{type}{control if the spline starts and ends at the terminal control points.}

\item{n}{number of points generated for the spline}

\item{orientation}{option to rotate the point set 90° clockwise}

\item{width, alpha}{arguments passed to \code{geom_bspline()}}

\item{coord}{coordinate system passed to \code{ggplot()}}
}
\value{
a ggplot object
}
\description{
Render a spline curve from control points
}
