% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_rendering.R
\name{theme_paper}
\alias{theme_paper}
\title{Customize a theme object with margins, color, and background.}
\usage{
theme_paper(
  output = "digital",
  color = "#FEFAEE",
  border = NULL,
  width = NULL,
  margin = rep(0, 4)
)
}
\arguments{
\item{output}{string to select "digital" or "plotter" theme.
"plotter" theme strips out panel backgroung and borders for a clean svg output.}

\item{color}{background color for digital output, default to an oldish white color.}

\item{border, width}{border color and width, default to NULL.}

\item{margin}{a vector of four numeric values for margins (top, right, bottom, left) in mm.}
}
\value{
a ggplot theme object
}
\description{
Customize a theme object with margins, color, and background.
}
