% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_rendering.R
\name{theme_paper}
\alias{theme_paper}
\title{Customize a theme object with margins, color, and background.}
\usage{
theme_paper(output = "digital", color = "#FEFAEE", margin = rep(0, 4))
}
\arguments{
\item{output}{set up background for "paper" or "plotter" output.}

\item{color}{background color for paper themes.}

\item{margin}{a vector of four numeric values for margins (top, right, bottom, left) in mm.}
}
\value{
a ggplot theme object
}
\description{
Customize a theme object with margins, color, and background.
}
