% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_geometry.R
\name{tr_rotate}
\alias{tr_rotate}
\title{Rotate a 2D object}
\usage{
tr_rotate(data, a, index = NULL)
}
\arguments{
\item{data}{dataframe with x and y coordinates, and id columns}

\item{a}{rotation angle (radian)}

\item{index}{string for the index column to be added from the original data}
}
\value{
a dataframe with new coordinates columns
}
\description{
Rotate a 2D object
}
