% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_geometry.R
\name{tr_rt}
\alias{tr_rt}
\title{Rotate then translate a 2D object}
\usage{
tr_rt(data, x0, y0, a, id = FALSE)
}
\arguments{
\item{data}{dataframe with x and y coordinates, and id columns}

\item{x0, y0}{shift in x and y axis (units)}

\item{a}{rotation angle (radian)}

\item{id}{if TRUE add the id column of the input to the result dataframe}
}
\value{
a dataframe with new coordinates columns
}
\description{
Rotate then translate a 2D object
}
