% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_geometry.R
\name{tr_translate}
\alias{tr_translate}
\title{Translate a 2D object}
\usage{
tr_translate(data, x0, y0, index = NULL)
}
\arguments{
\item{data}{dataframe with x and y coordinates, and id columns}

\item{x0, y0}{shift in x and y axis (units)}

\item{index}{string for the index column to be added from the original data}
}
\value{
a dataframe with new coordinates columns
}
\description{
Translate a 2D object
}
